+++ CHANGELOG.md
@@ -6,5 +6,12 @@ Release available for download on [GitHub](https://github.com/fmnext/maya/releas
 
 ## Release History
 
+## <small>2.7.1 (2026-06-30)</small>
+* changed api endpoints
+* fix unexpected FM23 tire behavior
+* add masks data to manufacturer colors json
+* fix missing tireR reference
+* fix thumbnail handler
+
 ## <small>2.7.0 (2026-06-01)</small>
 * Initial open source release